Add FastAPI chat endpoint and local server setup
