+++ README.md
@@ -18,7 +18,7 @@ Conversational CLI agent that embodies a non-clinical, autonomy-supportive physi
      ```
    - The application uses `python-dotenv` to load this file automatically.
 
-## Usage
+## Usage (CLI)
 
 Run the CLI:
 
@@ -33,7 +33,32 @@ Type messages to converse with the coach. `exit`/`quit` (or Ctrl+C/D) cleanly st
 - `main.py` – CLI entry point, conversation loop, OpenAI integration, and heuristic state inference.
 - `prompts/` – houses the base prompt and helper to inject inferred variables.
 - `requirements.txt` – Python dependencies.
+- `backend/` – FastAPI app + session store for streaming HTTP access.
+- `frontend/` – Static vanilla-JS chat experience that talks to the API.
 - `.env` – local secrets (never commit real keys).
+
+## Backend API + Web UI
+
+1. Ensure Docker Desktop is running (Qdrant dependency).
+2. Start the API server (this also handles virtualenv + dependencies):
+   ```bash
+   ./scripts/quickstart.sh api
+   ```
+   Override host/port with `API_HOST`/`API_PORT` if needed.
+3. Serve the static frontend (any static host works). During local dev you can run:
+   ```bash
+   cd frontend
+   python -m http.server 4173
+   ```
+   Then visit `http://localhost:4173` in your browser. The page expects the API at `http://localhost:8000` by default; set `window.API_BASE_URL` in `index.html` if you deploy elsewhere.
+
+The API exposes three endpoints:
+
+- `POST /sessions` → create anonymous session
+- `POST /sessions/{id}/messages` → stream assistant reply as newline-delimited JSON events
+- `DELETE /sessions/{id}` → clear a session (used by the “Start fresh” button)
+
+`GET /healthz` returns a simple readiness signal you can wire into uptime checks.
 
 ## Notes
