+++ coach/__init__.py
@@ -1,0 +1,5 @@
+"""Coach package exposing the conversational agent."""
+
+from .agent import CoachAgent, ConversationState, run_rag_sanity_check
+
+__all__ = ["CoachAgent", "ConversationState", "run_rag_sanity_check"]
