+++ requirements.txt
@@ -3,3 +3,5 @@ openai>=1.0.0
 llama-index
 qdrant-client
 llama-index-vector-stores-qdrant
+fastapi
+uvicorn[standard]
