+++ tests/test_generate_response.py
@@ -2,7 +2,7 @@
 
 import unittest
 
-from main import CoachAgent
+from coach import CoachAgent
 
 
 class StubCompletion:
