+++ tests/test_rag_routing.py
@@ -12,7 +12,7 @@ if str(PROJECT_ROOT) not in sys.path:
     sys.path.insert(0, str(PROJECT_ROOT))
 
 try:
-    from main import CoachAgent
+    from coach import CoachAgent
     from rag.retriever import RetrievedChunk, RetrievalResult
     from rag.router import QueryRouter, RouteDecision
